requirements.txt
numpy
torch 
torchvision
facenet-pytorch
ultralytics
opencv-python
streamlit
scikit-learn
pillow
matplotlib
seaborn
